chore: example upgrade to rn 0.74.3

+++ android/src/main/old-arch/com/rnmapbox/rnmbx/NativeRNMBXMovePointShapeAnimatorModuleSpec.java
@@ -36,9 +36,9 @@ public abstract class NativeRNMBXMovePointShapeAnimatorModuleSpec extends ReactC
 
   @ReactMethod
   @DoNotStrip
-  public abstract void create(int tag, ReadableArray coordinate, Promise promise);
+  public abstract void create(double tag, ReadableArray coordinate, Promise promise);
 
   @ReactMethod
   @DoNotStrip
-  public abstract void moveTo(int tag, ReadableArray coordinate, double duration, Promise promise);
+  public abstract void moveTo(double tag, ReadableArray coordinate, double duration, Promise promise);
 }
